// frontend/shared/src/office-tools/calculator/enums/calculator-key.ts
export enum CalculatorKey {
  AC = "AC",
  OpenBracket = "OpenBracket",
  CloseBracket = "CloseBracket",
  Reciprocal = "Reciprocal",
  Zero = "Zero",
  One = "One",
  Two = "Two",
  Three = "Three",
  Four = "Four",
  Five = "Five",
  Six = "Six",
  Seven = "Seven",
  Eight = "Eight",
  Nine = "Nine",
  Separator = "Separator",
  Divide = "Divide",
  Multiply = "Multiply",
  Subtraction = "Subtraction",
  Addition = "Addition",
  Exponentiation = "Exponentiation",
  Result = "Result"
}